+++ mayan/apps/document_states/tests/mixins/workflow_template_state_escalation_mixins.py
@@ -35,11 +35,11 @@ class WorkflowTemplateStateEscalationAPIViewTestMixin:
         )
 
         try:
-            self._test_workflow_estate_escalation = WorkflowStateEscalation.objects.get(
+            self._test_workflow_template_state_escalation = WorkflowStateEscalation.objects.get(
                 ~Q(pk__in=pk_list)
             )
         except WorkflowStateEscalation.DoesNotExist:
-            self._test_workflow_estate_escalation = None
+            self._test_workflow_template_state_escalation = None
 
         return response
 
@@ -61,27 +61,37 @@ class WorkflowTemplateStateEscalationAPIViewTestMixin:
             }
         )
 
-    def _request_test_workflow_template_state_escalation_edit_via_patch_api_view(self):
+    def _request_test_workflow_template_state_escalation_edit_via_patch_api_view(self, extra_data=None):
+        data = {
+            'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT_EDITED
+        }
+
+        if extra_data is not None:
+            data.update(extra_data)
+
         return self.patch(
             viewname='rest_api:workflow-template-state-escalation-detail', kwargs={
                 'workflow_template_id': self._test_workflow_template.pk,
                 'workflow_template_state_id': self._test_workflow_template_states[0].pk,
                 'workflow_template_state_escalation_id': self._test_workflow_template_state_escalation.pk
-            }, data={
-                'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT_EDITED
-            }
+            }, data=data
         )
 
-    def _request_test_workflow_template_state_escalation_edit_via_put_api_view(self):
+    def _request_test_workflow_template_state_escalation_edit_via_put_api_view(self, extra_data=None):
+        data = {
+            'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT_EDITED,
+            'workflow_template_transition_id': self._test_workflow_template_transitions[0].pk
+        }
+
+        if extra_data is not None:
+            data.update(extra_data)
+
         return self.put(
             viewname='rest_api:workflow-template-state-escalation-detail', kwargs={
                 'workflow_template_id': self._test_workflow_template.pk,
                 'workflow_template_state_id': self._test_workflow_template_states[0].pk,
                 'workflow_template_state_escalation_id': self._test_workflow_template_state_escalation.pk
-            }, data={
-                'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT_EDITED,
-                'workflow_template_transition_id': self._test_workflow_template_transitions[0].pk
-            }
+            }, data=data
         )
 
     def _request_test_workflow_template_state_escalation_list_api_view(self):
@@ -109,7 +119,7 @@ class WorkflowTemplateStateEscalationTaskTestMixin:
 
 
 class WorkflowTemplateStateEscalationTestMixin:
-    def _create_test_workflow_template_state_escalation(self, extra_data=None):
+    def _create_test_workflow_template_state_escalation(self, extra_kwargs=None):
         kwargs = {
             'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT,
             'comment': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_COMMENT,
@@ -117,9 +127,71 @@ class WorkflowTemplateStateEscalationTestMixin:
             'unit': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_UNIT
         }
 
-        if extra_data:
-            kwargs.update(extra_data)
+        if extra_kwargs:
+            kwargs.update(extra_kwargs)
 
         self._test_workflow_template_state_escalation = self._test_workflow_template_states[0].escalations.create(
             **kwargs
+        )
+
+
+class WorkflowTemplateStateEscalationViewTestMixin:
+    def _request_test_workflow_template_state_escalation_create_view(self, extra_data=None):
+        data = {
+            'transition': self._test_workflow_template_transition.pk
+        }
+
+        if extra_data is not None:
+            data.update(extra_data)
+
+        pk_list = list(
+            WorkflowStateEscalation.objects.values_list('pk', flat=True)
+        )
+
+        response = self.post(
+            viewname='document_states:workflow_template_state_escalation_create',
+            kwargs={
+                'workflow_template_state_id': self._test_workflow_template_states[0].pk
+            }, data=data
+        )
+
+        try:
+            self._test_workflow_template_state_escalation = WorkflowStateEscalation.objects.get(
+                ~Q(pk__in=pk_list)
+            )
+        except WorkflowStateEscalation.DoesNotExist:
+            self._test_workflow_template_state_escalation = None
+
+        return response
+
+    def _request_test_workflow_template_state_escalation_delete_view(self):
+        return self.post(
+            viewname='document_states:workflow_template_state_escalation_delete',
+            kwargs={
+                'workflow_template_state_escalation_id': self._test_workflow_template_state_escalation.pk
+            }
+        )
+
+    def _request_test_workflow_template_state_escalation_edit_view(self, extra_data=None):
+        data = {
+            'amount': TEST_WORKFLOW_TEMPLATE_STATE_ESCALATION_AMOUNT_EDITED,
+            'transition': self._test_workflow_template_state_escalation.transition.pk
+        }
+
+        if extra_data is not None:
+            data.update(extra_data)
+
+        return self.post(
+            viewname='document_states:workflow_template_state_escalation_edit',
+            kwargs={
+                'workflow_template_state_escalation_id': self._test_workflow_template_state_escalation.pk
+            }, data=data
+        )
+
+    def _request_test_workflow_template_state_escalation_list_view(self):
+        return self.get(
+            viewname='document_states:workflow_template_state_escalation_list',
+            kwargs={
+                'workflow_template_state_id': self._test_workflow_template_states[0].pk
+            }
         )