+++ mayan/apps/document_states/tests/test_workflow_instance_views.py
@@ -10,7 +10,7 @@ from .mixins.workflow_instance_mixins import WorkflowInstanceViewTestMixin
 from .mixins.workflow_template_mixins import (
     WorkflowTemplateTestMixin, WorkflowTemplateViewTestMixin
 )
-from .mixins.workflow_template_transition_mixins import WorkflowTransitionFieldTestMixin
+from .mixins.workflow_template_transition_field_mixins import WorkflowTemplateTransitionFieldTestMixin
 
 
 class WorkflowInstanceTransitionViewTestCase(
@@ -635,7 +635,7 @@ class WorkflowInstanceTransitionViewTestCase(
 
 
 class WorkflowInstanceTransitionFieldViewTestCase(
-    WorkflowTemplateTestMixin, WorkflowTransitionFieldTestMixin,
+    WorkflowTemplateTestMixin, WorkflowTemplateTransitionFieldTestMixin,
     WorkflowInstanceViewTestMixin, GenericDocumentViewTestCase
 ):
     auto_upload_test_document = False