+++ mayan/apps/document_states/tests/test_workflow_template_models.py
@@ -3,7 +3,9 @@ from mayan.apps.testing.tests.base import BaseTestCase
 from .mixins.workflow_template_mixins import WorkflowTemplateTestMixin
 
 
-class WorkflowModelTestCase(WorkflowTemplateTestMixin, BaseTestCase):
+class WorkflowTemplateModelTestCase(WorkflowTemplateTestMixin, BaseTestCase):
     def test_workflow_template_preview(self):
         self._create_test_workflow_template()
-        self.assertTrue(self._test_workflow_template.get_api_image_url())
+        self.assertTrue(
+            self._test_workflow_template.get_api_image_url()
+        )