+++ mayan/apps/document_states/tests/test_workflow_template_state_escalation_api.py
@@ -68,13 +68,44 @@ class WorkflowTemplateStateEscalationsAPIViewTestCase(
 
         self.assertEqual(events[0].actor, self._test_case_user)
         self.assertEqual(
-            events[0].action_object, self._test_workflow_estate_escalation
+            events[0].action_object, self._test_workflow_template_state_escalation
         )
         self.assertEqual(events[0].target, self._test_workflow_template)
         self.assertEqual(events[0].verb, event_workflow_template_edited.id)
 
     def test_workflow_template_state_escalation_create_api_view_validation_with_access(self):
         self._create_test_workflow_template_state_escalation()
+
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_template_edit
+        )
+
+        test_workflow_template_state_count = self._test_workflow_template_states[0].escalations.count()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_template_state_escalation_create_api_view()
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        self._test_workflow_template.refresh_from_db()
+        self.assertEqual(
+            self._test_workflow_template_states[0].escalations.count(),
+            test_workflow_template_state_count
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_template_state_escalation_create_api_view_transition_invalid_access(self):
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition(
+            extra_kwargs={
+                'origin_state': self._test_workflow_template_states[2],
+                'destination_state': self._test_workflow_template_states[3]
+            }
+        )
 
         self.grant_access(
             obj=self._test_workflow_template,
@@ -175,37 +206,6 @@ class WorkflowTemplateStateEscalationsAPIViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_workflow_template_state_escalation_list_api_view_no_permission(self):
-        self._create_test_workflow_template_state_escalation()
-
-        self._clear_events()
-
-        response = self._request_test_workflow_template_state_escalation_list_api_view()
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_workflow_template_state_escalation_list_api_view_with_access(self):
-        self._create_test_workflow_template_state_escalation()
-
-        self.grant_access(
-            obj=self._test_workflow_template,
-            permission=permission_workflow_template_view
-        )
-
-        self._clear_events()
-
-        response = self._request_test_workflow_template_state_escalation_list_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            response.data['results'][0]['id'],
-            self._test_workflow_template_state_escalation.pk
-        )
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
     def test_workflow_template_state_escalation_edit_api_view_via_patch_no_permission(self):
         self._create_test_workflow_template_state_escalation()
 
@@ -255,6 +255,43 @@ class WorkflowTemplateStateEscalationsAPIViewTestCase(
         )
         self.assertEqual(events[0].target, self._test_workflow_template)
         self.assertEqual(events[0].verb, event_workflow_template_edited.id)
+
+    def test_workflow_template_state_escalation_edit_api_view_via_patch_transition_invalid_with_access(self):
+        self._create_test_workflow_template_state_escalation()
+
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition(
+            extra_kwargs={
+                'origin_state': self._test_workflow_template_states[2],
+                'destination_state': self._test_workflow_template_states[3]
+            }
+        )
+
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_template_edit
+        )
+
+        test_workflow_template_state_escalation_amount = self._test_workflow_template_state_escalation.amount
+
+        self._clear_events()
+
+        response = self._request_test_workflow_template_state_escalation_edit_via_patch_api_view(
+            extra_data={
+                'workflow_template_transition_id': self._test_workflow_template_transition.pk
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        self._test_workflow_template_state_escalation.refresh_from_db()
+        self.assertEqual(
+            self._test_workflow_template_state_escalation.amount,
+            test_workflow_template_state_escalation_amount
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_workflow_template_state_escalation_edit_api_view_via_put_no_permission(self):
         self._create_test_workflow_template_state_escalation()
@@ -315,3 +352,76 @@ class WorkflowTemplateStateEscalationsAPIViewTestCase(
         )
         self.assertEqual(events[0].target, self._test_workflow_template)
         self.assertEqual(events[0].verb, event_workflow_template_edited.id)
+
+    def test_workflow_template_state_escalation_edit_api_view_via_put_transition_invalid_with_access(self):
+        self._create_test_workflow_template_state_escalation()
+
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition(
+            extra_kwargs={
+                'origin_state': self._test_workflow_template_states[2],
+                'destination_state': self._test_workflow_template_states[3]
+            }
+        )
+
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_template_edit
+        )
+
+        test_workflow_template_state_escalation_amount = self._test_workflow_template_state_escalation.amount
+        test_workflow_template_state_escalation_transition = self._test_workflow_template_state_escalation.transition
+
+        self._clear_events()
+
+        response = self._request_test_workflow_template_state_escalation_edit_via_put_api_view(
+            extra_data={
+                'workflow_template_transition_id': self._test_workflow_template_transition.pk
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        self._test_workflow_template_state_escalation.refresh_from_db()
+        self.assertEqual(
+            self._test_workflow_template_state_escalation.amount,
+            test_workflow_template_state_escalation_amount
+        )
+        self.assertEqual(
+            self._test_workflow_template_state_escalation.transition,
+            test_workflow_template_state_escalation_transition
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_template_state_escalation_list_api_view_no_permission(self):
+        self._create_test_workflow_template_state_escalation()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_template_state_escalation_list_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_template_state_escalation_list_api_view_with_access(self):
+        self._create_test_workflow_template_state_escalation()
+
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_template_view
+        )
+
+        self._clear_events()
+
+        response = self._request_test_workflow_template_state_escalation_list_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(
+            response.data['results'][0]['id'],
+            self._test_workflow_template_state_escalation.pk
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)