+++ mayan/apps/document_states/tests/test_workflow_template_transition_field_api.py
@@ -9,14 +9,16 @@ from ..permissions import (
 )
 
 from .mixins.workflow_template_mixins import WorkflowTemplateTestMixin
-from .mixins.workflow_template_transition_mixins import (
-    WorkflowTransitionFieldAPIViewTestMixin, WorkflowTransitionFieldTestMixin
+from .mixins.workflow_template_transition_field_mixins import (
+    WorkflowTemplateTransitionFieldAPIViewTestMixin,
+    WorkflowTemplateTransitionFieldTestMixin
 )
 
 
 class WorkflowTemplateTransitionFieldAPIViewTestCase(
-    WorkflowTransitionFieldAPIViewTestMixin, DocumentTestMixin,
-    WorkflowTemplateTestMixin, WorkflowTransitionFieldTestMixin, BaseAPITestCase
+    WorkflowTemplateTransitionFieldAPIViewTestMixin, DocumentTestMixin,
+    WorkflowTemplateTestMixin, WorkflowTemplateTransitionFieldTestMixin,
+    BaseAPITestCase
 ):
     auto_upload_test_document = False