+++ mayan/apps/document_states/tests/test_workflow_template_transition_views.py
@@ -1,7 +1,7 @@
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
 from ..events import event_workflow_template_edited
-from ..models import WorkflowTransition
+from ..models.workflow_transition_models import WorkflowTransition
 from ..permissions import (
     permission_workflow_template_edit, permission_workflow_template_view
 )
@@ -15,7 +15,7 @@ from .mixins.workflow_template_transition_mixins import (
 )
 
 
-class WorkflowTransitionViewTestCase(
+class WorkflowTemplateTransitionViewTestCase(
     WorkflowTemplateTestMixin, WorkflowTemplateViewTestMixin,
     WorkflowTemplateTransitionViewTestMixin, GenericViewTestCase
 ):